Career Mentor Agent App/requirements.txt
google-generativeai==0.3.2
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
plotly==5.17.0 